src/main/java/ru/geracimov/leetcode/top_interview_questions/easy/linkedlist/RemoveNthNodeFromEndOfList.java
package ru.geracimov.leetcode.top_interview_questions.easy.linkedlist;


/**
 * Given the head of a linked list, remove the nth node from the end of the list and return its head.
 */
public class RemoveNthNodeFromEndOfList {

    public ListNode removeNthFromEnd(ListNode head, int n) {
        ListNode prev = new ListNode(0);
        prev.next = head;
        ListNode first = prev;
        ListNode second = prev;
        for (int i = 1; i <= n + 1; i++) {
            first = first.next;
        }
        while (first != null) {
            first = first.next;
            second = second.next;
        }
        second.next = second.next.next;
        return prev.next;
    }

}
